% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{bin_counts}
\alias{bin_counts}
\title{Bin a counts data.table into a matrix of a given resolution}
\usage{
bin_counts(counts, resolution, b1 = NULL, b2 = NULL, e1 = NULL,
  e2 = NULL)
}
\arguments{
\item{counts}{data.table as returned by \code{\link{prepare_for_sparse_cs_norm}}}

\item{resolution}{positive integer.}

\item{b1, }{b2, e1, e2 Begins and ends of the portion of the data to bin. If NULL replace by min/max value.}
}
\value{
a data.table representing the binned data
}
\description{
Bin a counts data.table into a matrix of a given resolution
}

