% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{compute_gamma_overlap}
\alias{compute_gamma_overlap}
\title{compute \f$p(\Gamma_2>\Gamma_1) = \int_{0}^{+\infty} dx p_{\Gamma_2}(x) \int_{0}^{x} dy p_{\Gamma_1}(y)\f$}
\usage{
compute_gamma_overlap(alpha1, beta1, alpha2, beta2, bounds = 5, ncores = 1)
}
\description{
compute \f$p(\Gamma_2>\Gamma_1) = \int_{0}^{+\infty} dx p_{\Gamma_2}(x) \int_{0}^{x} dy p_{\Gamma_1}(y)\f$
}
\keyword{internal}

