% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{compute_normal_overlap}
\alias{compute_normal_overlap}
\title{compute \f$p(\mathcal{N}_2>\mathcal{N}_1) = \int_{-infty}^{+infty} dx p_{\mathcal{N}_2}(x) \int_{-infty}^{x} dy p_{\mathcal{N}_1}(y)\f$}
\usage{
compute_normal_overlap(mu1, sd1, mu2, sd2, bounds = 5, ncores = 1)
}
\description{
compute \f$p(\mathcal{N}_2>\mathcal{N}_1) = \int_{-infty}^{+infty} dx p_{\mathcal{N}_2}(x) \int_{-infty}^{x} dy p_{\mathcal{N}_1}(y)\f$
}
\keyword{internal}

