\name{csnorm-package}
\alias{csnorm-package}
\alias{csnorm}
\docType{package}
\title{
\packageTitle{csnorm}
}
\description{
\packageDescription{csnorm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{csnorm}
\packageIndices{csnorm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{csnorm}

Maintainer: \packageMaintainer{csnorm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
