% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{csnorm_fit}
\alias{csnorm_fit}
\title{Single-cpu fitting}
\usage{
csnorm_fit(model, biases, counts, dmin, dmax, bf_per_kb = 1,
  bf_per_decade = 5, iter = 10000, verbose = T, init = 0, ...)
}
\description{
Single-cpu fitting
}
\keyword{internal}

