% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{csnorm_fit_extradiag}
\alias{csnorm_fit_extradiag}
\title{Single-cpu fitting, only diagonal decay}
\usage{
csnorm_fit_extradiag(model, biases1, biases2, counts, dmin, dmax,
  bf_per_decade = 5, iter = 10000, verbose = T, init = 0, ...)
}
\description{
Single-cpu fitting, only diagonal decay
}
\keyword{internal}

