% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{csnorm_predict_all}
\alias{csnorm_predict_all}
\title{Predict expected values for each count given optimized model parameters}
\usage{
csnorm_predict_all(model, biases, counts, opt, bf_per_decade = 5,
  verbose = T)
}
\arguments{
\item{model}{the stan model}

\item{biases, }{counts as returned by \code{\link{prepare_for_sparse_cs_norm}} or a subset of it}

\item{opt}{the optimized parameters, as returned by \code{\link{run_split_parallel}}}

\item{bf_per_decade}{same as that used in \code{\link{run_split_parallel}}}

\item{verbose}{}
}
\value{
a stan optimization output with the predictions.
}
\description{
Predict expected values for each count given optimized model parameters
}
\keyword{internal}

