% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{csnorm_predict_binned}
\alias{csnorm_predict_binned}
\title{Bin observed and expected counts at a given resolution}
\usage{
csnorm_predict_binned(model, biases, counts, opt, resolution, b1 = NULL,
  b2 = NULL, e1 = NULL, e2 = NULL, bf_per_decade = 5, verbose = F,
  circularize = -1L)
}
\description{
Bin observed and expected counts at a given resolution
}
\keyword{internal}

