% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{detect_interactions}
\alias{detect_interactions}
\title{Detect significant interactions wrt expected}
\usage{
detect_interactions(binned, dispersions, threshold = 0.95, ncores = 1,
  normal.approx = 100)
}
\arguments{
\item{binned}{as returned by \code{\link{csnorm_predict_binned}}}

\item{dispersions}{as returned by \code{\link{get_dispersions}}}

\item{threshold}{significance threshold, between 0 and 1}

\item{ncores}{number of cores used for parallelization}

\item{normal.approx}{use normal approximation if dispersion is larger than this}
}
\value{
the binned matrix with additional information relating to these significant interactions
}
\description{
Detect significant interactions wrt expected
}
\keyword{internal}

