% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{diagnose_biases}
\alias{diagnose_biases}
\title{Plot bias run statistics using intermediate files}
\usage{
diagnose_biases(outprefix, coverage = 4, square.size = 150000)
}
\description{
Plot bias run statistics using intermediate files
}
\keyword{internal}

