% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{diagnose_counts}
\alias{diagnose_counts}
\title{Plot bias run statistics using intermediate files}
\usage{
diagnose_counts(outprefix, coverage.extradiag = 1, square.size = 150000)
}
\description{
Plot bias run statistics using intermediate files
}
\keyword{internal}

