% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{dset_statistics}
\alias{dset_statistics}
\title{Print general statistics on the output of \code{\link{prepare_for_sparse_cs_norm}}}
\usage{
dset_statistics(biases, counts)
}
\arguments{
\item{biases, }{counts}
}
\description{
Print general statistics on the output of \code{\link{prepare_for_sparse_cs_norm}}
}

