% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{fill_zeros}
\alias{fill_zeros}
\title{Convert a sparse counts data.table to a dense one by adding rows with zero counts}
\usage{
fill_zeros(counts, biases, biases2 = NULL)
}
\arguments{
\item{counts, }{biases data.tables as returned by \code{\link{prepare_for_sparse_cs_norm}}}

\item{biases2}{data.table of biases for id2 column of counts. If NULL (default), use that of biases}
}
\value{
a counts data.table with zeros filled according to cut sites provided in biases (and biases2 if available)
}
\description{
Convert a sparse counts data.table to a dense one by adding rows with zero counts
}
\section{Warning}{

Memory-intensive
}
\keyword{internal}

