% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{generate_fake_dataset}
\alias{generate_fake_dataset}
\title{Generate a fake dataset that follows the cut site normalization model's specifications}
\usage{
generate_fake_dataset(num_rsites = 10, genome_size = 10000, eC = 0.1,
  eRJ = 0.4, eDE = 0.8, alpha = 10)
}
\arguments{
\item{num_rsites}{Number of restriction sites}

\item{genome_size}{Size of the genome}

\item{eC, }{eRJ, eDE Exposure for counts, rejoined and dangling ends}

\item{alpha}{Dispersion}
}
\value{
a list of biases and counts, similar to \code{\link{prepare_for_sparse_cs_norm}}
}
\description{
Generate a fake dataset that follows the cut site normalization model's specifications
}

