% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_cs_subset}
\alias{get_cs_subset}
\title{Get subset of counts and biases data.table}
\usage{
get_cs_subset(counts, biases, begin1, end1, begin2 = NULL, end2 = NULL,
  fill.zeros = T, circularize = -1)
}
\arguments{
\item{counts, }{biases as returned by \code{\link{prepare_for_sparse_cs_norm}}}

\item{begin1, }{end1 integers. Genomic location of begin/end of subset to take}

\item{begin2, }{end2 integers. If provided, take extradiagonal section}

\item{fill.zeros}{whether to return a dense matrix (default TRUE)}

\item{integer.}{Set this to the size of the chromosome if it is
circular, otherwise leave as-is (default is -1)}
}
\value{
counts and biases data.tables, in a list
}
\description{
Get subset of counts and biases data.table
}
\section{Warning}{

Memory-intensive
}

