% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_dispersions}
\alias{get_dispersions}
\title{Predict dispersions for a binned matrix}
\usage{
get_dispersions(model, binned, iter = 10000, verbose = T)
}
\description{
Predict dispersions for a binned matrix
}
\keyword{internal}

