% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_subset}
\alias{get_subset}
\title{Get subset of reads}
\usage{
get_subset(data, b1, e1, b2 = NULL, e2 = NULL)
}
\arguments{
\item{data}{}

\item{b1, }{e1 begin / end of read range you want}

\item{b2, }{e2 (optional) if provided, give extradiagonal portion}
}
\value{
data.table
}
\description{
Get subset of reads
}
\keyword{internal}

