% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{iterative_normalization}
\alias{iterative_normalization}
\title{Apply the ICE algorithm to a binned matrix}
\usage{
iterative_normalization(bdata, niterations = 100)
}
\arguments{
\item{bdata}{binned data.table as returned by \code{\link{bin_counts}}}

\item{niterations}{positive integer. Number of iterations to perform}
}
\value{
the ICEd matrix
}
\description{
Apply the ICE algorithm to a binned matrix
}

