% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Wrapper for the postprocessing steps}
\usage{
postprocess(biases, counts, op, resolution = 10000, ncores = 30,
  predict.all.means = T, circularize = -1L)
}
\arguments{
\item{biases}{}

\item{counts}{}

\item{op}{}

\item{resolution}{}

\item{ncores}{}

\item{predict.all.means}{}

\item{circularize}{}
}
\description{
Wrapper for the postprocessing steps
}

