% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_for_sparse_cs_norm}
\alias{prepare_for_sparse_cs_norm}
\title{Parse reads data.table and extract necessary objects for cut-site
normalization}
\usage{
prepare_for_sparse_cs_norm(data, both = F, circularize = -1)
}
\arguments{
\item{data}{reads data.table}

\item{both}{boolean. Whether a merged table providing biases and their counts
should be produced (default is FALSE as it is not used by cs norm)}

\item{circularize}{integer. Set this to the size of the chromosome if it is
circular, otherwise leave as-is (default is -1)}
}
\value{
A list containing biases and counts, two data tables representing the
  total number of reads of different categories at each cut site.
}
\description{
Parse reads data.table and extract necessary objects for cut-site
normalization
}

