% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_and_prepare}
\alias{read_and_prepare}
\title{Wrapper function for the whole preprocessing}
\usage{
read_and_prepare(infile, outprefix, skip = 0L, both = T,
  distance_bins_per_decade = 100, circularize = -1, dangling.L = c(0, 4),
  dangling.R = c(3, -1))
}
\arguments{
\item{infile}{}

\item{outprefix}{}

\item{skip}{}

\item{both}{}

\item{distance_bins_per_decade}{}

\item{circularize}{}

\item{dangling.L}{}

\item{dangling.R}{}
}
\description{
Wrapper function for the whole preprocessing
}

