% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_biases}
\alias{run_split_parallel_biases}
\title{Genomic bias estimation part of parallel run}
\usage{
run_split_parallel_biases(smfit, counts, biases, begin, end, dmin, dmax,
  bf_per_kb, bf_per_decade, verbose, iter, outprefix = NULL,
  circularize = -1)
}
\description{
Genomic bias estimation part of parallel run
}
\keyword{internal}

