% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_biases_homogenize}
\alias{run_split_parallel_biases_homogenize}
\title{Homogenize biases estimated on multiple fits}
\usage{
run_split_parallel_biases_homogenize(model, dt, lambda_nu, lambda_delta,
  bf_per_kb = 1, iter = 10000, verbose = T, init = 0, ...)
}
\description{
Homogenize biases estimated on multiple fits
}
\keyword{internal}

