% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_counts}
\alias{run_split_parallel_counts}
\title{Decay bias estimation part of parallel run}
\usage{
run_split_parallel_counts(model, counts, biases.aug, begin1, end1, begin2, end2,
  dmin, dmax, bf_per_decade = 5, verbose = T, iter = 1e+05,
  outprefix = NULL, circularize = -1L)
}
\description{
Decay bias estimation part of parallel run
}
\keyword{internal}

