% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_counts_decay}
\alias{run_split_parallel_counts_decay}
\title{Merge together parallel runs and estimate a single diagonal decay}
\usage{
run_split_parallel_counts_decay(model, dt, counts, dmin, dmax,
  bf_per_decade = 5, iter = 10000, verbose = T, init = 0, ...)
}
\description{
Merge together parallel runs and estimate a single diagonal decay
}
\keyword{internal}

