% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_counts_eC}
\alias{run_split_parallel_counts_eC}
\title{Get a proper count exposure for all estimated biases}
\usage{
run_split_parallel_counts_eC(model, biases, counts, retlist, dmin, dmax,
  bf_per_decade = 5, verbose = T, iter = 1000)
}
\description{
Get a proper count exposure for all estimated biases
}
\keyword{internal}

