% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_recovery}
\alias{run_split_parallel_recovery}
\title{Parse intermediate outputs of killed run and try to compute the estimates nevertheless}
\usage{
run_split_parallel_recovery(counts, biases, outprefix, square.size = 1e+05,
  coverage = 4, coverage.extradiag = 1, bf_per_kb = 1,
  bf_per_decade = 5, distance_bins_per_decade = 100, verbose = F,
  iter = 1e+05, ncpus = 30, homogenize = F, circularize = -1L)
}
\description{
Parse intermediate outputs of killed run and try to compute the estimates nevertheless
}
\keyword{internal}

