% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_split_parallel_squares}
\alias{run_split_parallel_squares}
\title{Split a chromosome into subsets for parallelization}
\usage{
run_split_parallel_squares(biases, square.size, coverage, diag.only = F)
}
\description{
Split a chromosome into subsets for parallelization
}
\keyword{internal}

