% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{thresholds_estimator}
\alias{thresholds_estimator}
\title{estimates the values of the count or dispersion required to cross a given threshold}
\usage{
thresholds_estimator(observed, expected, dispersion, threshold = 0.95,
  counts.range = c(expected/100, expected * 100), disp.range = c(0, 100 *
  dispersion), compute = T)
}
\arguments{
\item{observed, }{expected, dispersion, threshold floats used to generate the plots}

\item{counts.range}{scan counts in that range}

\item{disp.range}{scan dispersion in that range}

\item{compute}{whether to compute the values that meet the threshold or not. This step can fail if the ranges are set badly.}
}
\value{
a list of three plots
}
\description{
For illustration purposes. Has a border effect at high counts that can be safely ignored.
}

